--- nexcart/packages/nexcart-design-system/.storybook/Button/Button.mdx
@@ -1,50 +1,0 @@
-import {Canvas, Controls, Meta, Primary, Stories} from "@storybook/addon-docs";
-import * as ButtonStories from "./Button.stories";
-
-<Meta of={ButtonStories}/>
-
-# Button
-
-The Button component is a foundational UI element used for triggering actions or navigation.
-
-## Usage
-
-To use the button, import it from the design system:
-
-```tsx
-import {Button} from "@nexcart/nexcart-design-system";
-
-const MyComponent = () => (
-  <>
-    <Button>My button</Button>
-    <Button variant="secondary">My button</Button>
-  </>
-);
-```
-
-## Playground
-
-Interact with the button using the controls below to see how it responds to different props:
-
-<Canvas of={ButtonStories.Playground}/>
-
-<Controls of={ButtonStories.Playground}/>
-
-## Props
-
-- **variant**: `"primary" | "secondary"` - The visual style of the button (default: `undefined`)
-- **loading**: `boolean` - Shows the loading state and disables the button
-- **disabled**: `boolean` - Disables the button
-- **children**: `ReactNode` - The content inside the button
-
-All other standard HTML button attributes are supported and will be passed directly to the underlying `<button>` element.
-
-## All Variants
-
-Here's an overview of all button variants and states:
-
-<Canvas of={ButtonStories.All} layout="fullscreen"/>
-
-## All Stories
-
-<Stories/>
